// frontend/src/openapi/schema.ts
export interface components {
  schemas: {
    DocVector: number[];
    Question: {
      /** Format: uuid */
      id: string;
      summary: string;
      /** Format: uuid */
      userId: string;
      /** Format: date-time */
      finishedAt?: string;
      docVector: components["schemas"]["DocVector"];
      tags: components["schemas"]["Tag"][];
      topics: components["schemas"]["Topic"][];
      comments: components["schemas"]["QuestionComment"][];
      /** Format: date-time */
      createdAt: string;
    };
    Topic: {
      /** Format: uuid */
      id: string;
      /** Format: uuid */
      questionId?: string;
      title: string;
      description: string;
      isDefaultTopic: boolean;
      comments: components["schemas"]["TopicComment"][];
      /** Format: date-time */
      createdAt: string;
    };
    QuestionComment: {
      /** Format: uuid */
      id: string;
      /** Format: uuid */
      questionId: string;
      /** Format: uuid */
      userId: string;
      /** @enum {string} */
      type: "suggest" | "question" | "answer";
      content: string;
      /** Format: date-time */
      createdAt: string;
    };
    TopicComment: {
      /** Format: uuid */
      id: string;
      /** Format: uuid */
      topicId: string;
      /** Format: uuid */
      userId: string;
      /** @enum {string} */
      type: "suggest" | "question" | "answer";
      content: string;
      /** Format: date-time */
      createdAt: string;
    };
    Tag: {
      /** Format: uuid */
      id: string;
      name: string;
    };
    OriginalUser: {
      /** Format: uuid */
      id: string;
      name: string;
      photoUrl: string;
    };
    QuestionResponse: {
      /** Format: uuid */
      questionId?: string;
      status?: number;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}